Move extra attributes to init
